+++ config/variables.py
@@ -21,4 +21,4 @@ ALERTING_SEVERITIES = ['medium', 'high', 'critical']
 ROOT_PATH = Path(__file__).parent.parent
 FINDING_RULES_PATH = ROOT_PATH / 'config' / 'rules.yaml'
 
-META_TYPES = ['EC2', 'ELBV2', 'RDS', 'S3']
+META_TYPES = ['EC2', 'ELBV2', 'IAM', 'RDS', 'S3']